requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-deepseek==0.0.1
pyyaml==6.0.1
requests==2.31.0
pytest==7.4.3
pytest-html==4.1.1
jsonschema==4.20.0